24w24c: Mark RegistereduUserController as unused.
