+++ resources/views/user/index.blade.php
@@ -1,6 +1,5 @@
 @include('header');
 
 <p>List of users</p>
-<p>{{$users}}</p>
 
 @include('footer');
