+++ routes/web.php
@@ -36,8 +36,8 @@ Route::name('submit')->/*middleware('check.age')->*/group(function(){
 
 /*Route::get('/user/blogs/{blogTitle?}')*/
 
-Route::get('user/login', 'App\Http\Controllers\LoginController@show')->name('user.login');
-Route::post('user/login/config', 'App\Http\Controllers\LoginController@login')->name('user.login_config');
+Route::get('user/login', 'App\Http\Controllers\UserController@showLogin')->name('user.login');
+Route::post('user/login/config', 'App\Http\Controllers\UserController@login')->name('user.login_config');
  
 /* RESOURCES */
 //Route::resource('user', 'App\Http\Controllers\UserController');
@@ -45,8 +45,9 @@ Route::resource('post', 'App\Http\Controllers\PostController', ['parameters' => 
 Route::resource('article', 'App\Http\Controllers\ArticleController', ['only'=>['create', 'destroy', 'update', 'edit', 'store']]);
 
 
-Route::get('user/create', 'App\Http\Controllers\RegisterController@show')->name('user.create');
-Route::post('user/submit', 'App\Http\Controllers\RegisterController@register')->name('user.submit');
+Route::get('user/create', 'App\Http\Controllers\UserController@create')->name('user.create');
+Route::post('user/submit', 'App\Http\Controllers\UserController@store')->name('user.submit');
+Route::get('user/index', 'App\Http\Controllers\UserController@index')->name('user.index');
 
 
 /* REDIRECTS */
